지갑 연동 보강 v0.0.6.1

+++ src/account.rs
@@ -77,7 +77,7 @@ impl AccountWallet {
         hex::encode(bytes)
     }
 
-    pub fn sign_transfer(&self, to: Address, amount: u64, fee: u64, nonce: u64) -> Transaction {
+    pub fn sign_transfer(&self, to: Address, amount: u128, fee: u128, nonce: u64) -> Transaction {
         let mut tx = Transaction {
             from: self.address(),
             to,
